Allow running play.py without having scikit-learn and h5py installed.
