+++ play.py
@@ -1,10 +1,9 @@
-import train
+import load
 import pickle
 import theano
 import theano.tensor as T
 import math
 import chess, chess.pgn
-from parse_game import bb2array
 import heapq
 import time
 import re
@@ -20,8 +19,8 @@ def get_model_from_pickle(fn):
     f = open(fn)
     Ws, bs = pickle.load(f)
     
-    Ws_s, bs_s = train.get_parameters(Ws=Ws, bs=bs)
-    x, p = train.get_model(Ws_s, bs_s)
+    Ws_s, bs_s = load.get_parameters(Ws=Ws, bs=bs)
+    x, p = load.get_model(Ws_s, bs_s)
     
     predict = theano.function(
         inputs=[x],
